navigationMenu 함수 삭제 후 return 내부에서 바로 적용
